refactor: move business logic from prisma extensions to services

+++ src/services/ListService.ts
@@ -1,4 +1,5 @@
 import prisma from "@/prisma";
+import { DatabaseLimitError } from "@/errors/DatabaseLimitError";
 
 export class ListService {
   public static async listForUser(userId: number) {
@@ -9,6 +10,13 @@ export class ListService {
   }
 
   public static async create(userId: number, data: any) {
+    const amount = await prisma.list.count({
+      where: { userId },
+    });
+    if (amount >= 10) {
+      throw new DatabaseLimitError("Je mag maximaal 10 lijsten hebben.");
+    }
+
     return await prisma.list.create({
       data: {
         ...data,
@@ -25,6 +33,13 @@ export class ListService {
   }
 
   public static async delete(userId: number, listId: number) {
+    const amount = await prisma.list.count({
+      where: { userId },
+    });
+    if (amount <= 1) {
+      throw new DatabaseLimitError("Je moet minimaal 1 lijst hebben.");
+    }
+
     return await prisma.list.delete({ where: { id: listId, userId } });
   }
 }
